requirements.txt
aioredis==1.3.1
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
blessed==1.20.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-auth-token-middlewares==1.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
cloudinary==1.39.1
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Django==4.2.5
django-autoslug==1.9.9
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==24.2
django-picklefield==3.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
firebase-admin==6.5.0
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==21.2.0
hiredis==2.3.2
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
jinxed==1.2.1
msgpack==1.0.8
packaging==24.0
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytils==0.4.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
requests==2.31.0
rsa==4.9
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
zope.interface==6.4.post2
